+++ tomcat/src/main/java/nextstep/jwp/controller/IndexController.java
@@ -5,9 +5,10 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import nextstep.jwp.exception.InvalidRequestMethodException;
-import nextstep.jwp.http.common.ContentType;
-import nextstep.jwp.http.request.HttpRequest;
-import nextstep.jwp.http.response.HttpResponse;
+import org.apache.catalina.AbstractController;
+import org.apache.coyote.http11.common.ContentType;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public class IndexController extends AbstractController {
