+++ tomcat/src/main/java/nextstep/jwp/exception/NotFoundException.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.exception;
 
-import nextstep.jwp.http.common.HttpStatus;
+import org.apache.coyote.http11.common.HttpStatus;
 
 public class NotFoundException extends HttpGlobalException {
