+++ tomcat/src/main/java/org/apache/coyote/http11/common/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http.common;
+package org.apache.coyote.http11.common;
 
 public enum ContentType {
     TEXT_HTML("text/html;charset=utf-8"),
