+++ tomcat/src/main/java/org/apache/coyote/http11/common/Cookie.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http.common;
+package org.apache.coyote.http11.common;
 
 public class Cookie {
