+++ tomcat/src/main/java/org/apache/coyote/http11/common/HttpStatus.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http.common;
+package org.apache.coyote.http11.common;
 
 public enum HttpStatus {
     OK("200", "OK"),
