+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -1,8 +1,8 @@
-package nextstep.jwp.http.request;
+package org.apache.coyote.http11.request;
 
-import nextstep.jwp.http.common.Cookie;
-import nextstep.jwp.http.common.HttpBody;
-import nextstep.jwp.http.common.HttpVersion;
+import org.apache.coyote.http11.common.Cookie;
+import org.apache.coyote.http11.common.HttpBody;
+import org.apache.coyote.http11.common.HttpVersion;
 
 public class HttpRequest {
